Use separate bool variables to show each step in the app to avoid a WinUI bug with IsVisible
To apply the styles correctly in WinUI (i.e. in Windows builds), you have to make the whole page visible on initialization right now (this seems to be a bug within .NET MAUI).

Thus, we use a separate bool variable for each step to initialize each bool variable with true on initialization.

When the page is loaded, the Restart() method will be called which then sets only one of the boolean variables on true.

The idea for the solution is based on https://github.com/dotnet/maui/issues/11427. Wrapping the panels in another panel did not work.